Select option disabled and error registration handled
Dealt with the Activity instance. Created a default value for it isFull (like a virtual field) with settler and getter that I then use to push into the instance and show the select otpion disabled if the room is full.
Also checked in the '/ajax/processStep2' call if the user was already registered to the activity. If yes, I just show a warning message in the #Step2 in the view.